[hotfix] fix bug in getting user history in SeqDataset
**Problem Description**
The SeqDataset will not load the first item for each user as the user history, which is used for masking interacted items in get topk predictions. The reason lies in that the `inter_feat_subset` is set as the -1 dim of `data_index`, where only those items can be used as labels in training are returned, thereby ignoring the first item of each user (the first item will never be regarded as a label in training).

**How to fix**
To fix this issue, we need to modify the `inter_feat_subset` to include the first item of each user. We can do this by adding the first item index to the `inter_feat_subset` for each user.